meter per second works

+++ AirTrafficMonitoring/TOS/TOS.cs
@@ -13,8 +13,8 @@ namespace TOS
         int PosistionY { get;}
         int Altitude { get; }
         DateTime TimeStamp { get;}
-        int Velocity { get; set; }
-        int degress { get; set; }
+        double Velocity { get; set; }
+        double degress { get; set; }
 
     }
     public class TOS : ITOS
@@ -24,8 +24,8 @@ namespace TOS
         public int PosistionY { get; private set; }
         public int Altitude { get; private set; }
         public DateTime TimeStamp { get;private set; }
-        public int Velocity { get; set; }
-        public int degress { get; set; }
+        public double Velocity { get; set; }
+        public double degress { get; set; }
 
         public TOS(string tag, int posistionX, int posistionY, int altitude, DateTime timeStamp)
         {
